Updated stack with current work

+++ STV-Stack/src/Types.hs
@@ -4,4 +4,4 @@ type Candidates = [String]
 type Vote = (String, String)
 type Votes = [Vote]
 type SortedVotes = [[String]]
-type VotesRecieved = (String, Int)
+type VotesRecieved = (String, Double)
